Moved downloading boxoban levels to a new separate script
